+++ test/Playlist1Test.java
@@ -1,0 +1,3 @@
+public class Playlist1Test extends PlaylistTest {
+
+}