+++ CHANGELOG.md
@@ -6,6 +6,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Calendar Versioning](https://calver.org/) of
 the following form: YYYY.0M.0D.
 
+
+## 2025.12.05
+
+### Updated
+
+- Added standard kernel method implementations
+
 ## 2025.11.21
 
 ### Added
@@ -52,3 +59,18 @@ the following form: YYYY.0M.0D.
 - Designed a music playlist component
 - Designed a trick tracker component
 - Designed a photo library component
+
+## 2025.10.25
+
+### Added
+
+- Designed kernel and enhanced interfaces for playlist component
+
+### Updated
+
+- Changed design to include playlistKernel and playlist interfaces
+- Shuffle and startFrom methods introduced.
+- Added Song subclass to PlaylistKernel
+- Changed method headers in the interfaces to change from generic song type to song
+- Changed method headers in the playlist enhanced interface to treat Queue<Pair<S, A>> as this.
+.